create out_dir.spec.ts to divide cli.spec
Because I want to minify spec to speed up the test that i want to fix.
