+++ __tests__/cli.spec.ts
@@ -18,46 +18,3 @@ describe('confirm flags and options of cli', () => {
     }
   });
 });
-
-describe('compress png images', () => {
-  beforeAll(() => {
-    child_process.execSync('rm -rf tmp && mkdir tmp');
-  });
-
-  test('output the image to the path that --out-dir option set', () => {
-    const rawImage = fs.readFileSync('images/illust.png');
-    child_process.execSync('bin/cli images/illust.png --out-dir tmp');
-    const compressedImage = fs.readFileSync('tmp/illust.png');
-    expect(rawImage.byteLength).toBeGreaterThan(compressedImage.byteLength);
-  });
-});
-
-describe('compress jpg images', () => {
-  beforeAll(() => {
-    child_process.execSync('rm -rf tmp && mkdir tmp');
-  });
-
-  test('output the image to the path that --out-dir option set', () => {
-    const rawImage = fs.readFileSync('images/illust.png');
-    child_process.execSync('bin/cli images/illust.png --out-dir ./tmp');
-    const compressedImage = fs.readFileSync('tmp/illust.png');
-    expect(rawImage.byteLength).toBeGreaterThan(compressedImage.byteLength);
-  });
-
-  describe('compress images with regexp', () => {
-    beforeAll(() => {
-      child_process.execSync('rm -rf tmp && mkdir tmp');
-    });
-
-    test('output some images to tmp directory', () => {
-      const rawJpgImage = fs.readFileSync('images/sample.jpg');
-      const rawPngImage = fs.readFileSync('images/illust.png');
-      child_process.execSync('bin/cli images/* --out-dir ./tmp/');
-      const compressedJpgImage = fs.readFileSync('tmp/sample.jpg');
-      const compressedPngImage = fs.readFileSync('tmp/illust.png');
-
-      expect(rawJpgImage.byteLength).toBeGreaterThan(compressedJpgImage.byteLength);
-      expect(rawPngImage.byteLength).toBeGreaterThan(compressedPngImage.byteLength);
-    });
-  });
-});
